// angular-biometric-designer/src/app/shared/models/identity/role.dto.ts
export class RoleDto {
    id: string = '';
    id_trans: string = '';
    role_code: string = '';
    role_name: string = '';
    status: string = '';
    auth_stat: string = '';
    created_by: string = '';
    created_date: string = '';
    last_modified_by: string = '';
    last_modified_date: string = '';
    approved_by: string = '';
    approved_date: string = '';
}